@@ -1,24 +1,12 @@
-"use strict";
+import React from "react";
 
-var React = require('react');
+// Function component - Builds a basic navbar with anchors for pages
+function Header() {
+  return (
+    <nav>
+      <a href="/">Home</a> | <a href="/">About</a>
+    </nav>
+  );
+}
 
-var Header = React.createClass({
-    render: function () {
-        return (
-            <nav className="navbar navbar-default">
-                <div className="container-fluid">
-                    <a href="/" className="navbar-brand">
-                        <img src="images/pluralsight-logo.png" />
-                    </a>
-                    <ul className="nav navbar-nav">
-                        <li><a href="/">Home</a></li>
-                        <li><a href="/#authors">Authors</a></li>
-                        <li><a href="/#about">About</a></li>
-                    </ul>
-                </div>
-            </nav>
-        );
-    }
-});
-
-module.exports = Header;
+export default Header;
